test(ui): decompose dashboard header contract
